requirements.txt
# Django Core
django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Authentication
djangorestframework-simplejwt>=5.2.0
django-filter>=23.0

# Database
psycopg2-binary>=2.9.0  # PostgreSQL

# Cache
redis>=4.5.0
django-redis>=5.3.0

# Monitoring & Logging
sentry-sdk>=1.20.0
prometheus-client>=0.16.0
django-prometheus>=2.3.0

# Environment
django-environ>=0.10.0

# API Documentation
drf-spectacular>=0.26.0

# Async Support
channels>=4.0.0
channels-redis>=4.1.0

# Testing
pytest>=7.3.0
pytest-django>=4.5.0
factory-boy>=3.2.0
faker>=18.0.0
coverage>=7.2.0

# Development
ipython>=8.12.0
django-debug-toolbar>=4.0.0

# Production
gunicorn>=20.1.0
uvicorn[standard]>=0.22.0
whitenoise>=6.4.0